+++ CHANGES.rst
@@ -10,6 +10,8 @@ Bugfixes:
 * Fix type annotation for `AIOKafkaAdminClient` (issue #1148)
 * Return back and deprecate `api_version` parameter in client classes
   (issue #1147)
+* Avoid failures when a transaction coordinator is dead by removing deprecated code
+  (issue #1151)
 
 
 0.13.0 (2026-01-02)